Stop using servlet filter to secure api. Using session checks in each method instead

+++ src/main/java/com/practo/jedi/carpool/controller/AddressController.java
@@ -1,5 +1,8 @@
 package com.practo.jedi.carpool.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,32 +24,52 @@ public class AddressController {
   private AddressService service;
 
   @RequestMapping(method = RequestMethod.GET)
-  public Iterable<AddressModel> list() {
+  public Iterable<AddressModel> list(HttpSession session, HttpServletResponse servletResponse) {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get();
   }
 
   @RequestMapping(method = RequestMethod.POST)
-  public ResponseEntity<AddressModel> create(@RequestBody AddressModel address) throws EntityNotFoundException {
+  public ResponseEntity<AddressModel> create(@RequestBody AddressModel address, HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     AddressModel m = service.create(address);
     ResponseEntity<AddressModel> response = new ResponseEntity<AddressModel>(m, HttpStatus.CREATED);
     return response;
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  public AddressModel get(@PathVariable("id") int id) throws EntityNotFoundException {
+  public AddressModel get(@PathVariable("id") int id, HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get(id);
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
   public ResponseEntity<AddressModel> update(@PathVariable("id") int id,
-      @RequestBody AddressModel address) throws EntityNotFoundException {
+      @RequestBody AddressModel address, HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     AddressModel m = service.update(address, id);
     ResponseEntity<AddressModel> response = new ResponseEntity<AddressModel>(m, HttpStatus.OK);
     return response;
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-  public ResponseEntity<Boolean> delete(@PathVariable("id") int id) throws EntityNotFoundException {
+  public ResponseEntity<Boolean> delete(@PathVariable("id") int id, HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     service.delete(id);
     ResponseEntity<Boolean> response = new ResponseEntity<Boolean>(true, HttpStatus.NO_CONTENT);
     return response;
