+++ src/main/java/com/practo/jedi/carpool/controller/BookingController.java
@@ -1,5 +1,8 @@
 package com.practo.jedi.carpool.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,26 +24,46 @@ public class BookingController {
   private BookingService service;
 
   @RequestMapping(method = RequestMethod.GET)
-  public Iterable<BookingModel> list(@PathVariable("listing_id") int listing_id) {
+  public Iterable<BookingModel> list(@PathVariable("listing_id") int listing_id,
+      HttpSession session, HttpServletResponse servletResponse) {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get(listing_id);
   }
 
   @RequestMapping(method = RequestMethod.POST)
   public ResponseEntity<BookingModel> create(@PathVariable("listing_id") int listing_id,
-      @RequestBody BookingModel booking) throws EntityNotFoundException {
+      @RequestBody BookingModel booking, HttpSession session, HttpServletResponse servletResponse)
+      throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     BookingModel m = service.create(listing_id, booking);
     ResponseEntity<BookingModel> response = new ResponseEntity<BookingModel>(m, HttpStatus.CREATED);
     return response;
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  public BookingModel get(@PathVariable("listing_id") int listing_id, @PathVariable("id") int id) throws EntityNotFoundException {
+  public BookingModel get(@PathVariable("listing_id") int listing_id, @PathVariable("id") int id,
+      HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get(listing_id, id);
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
   public ResponseEntity<BookingModel> update(@PathVariable("listing_id") int listing_id,
-      @PathVariable("id") int id, @RequestBody BookingModel booking) throws EntityNotFoundException {
+      @PathVariable("id") int id, @RequestBody BookingModel booking, HttpSession session,
+      HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     BookingModel m = service.update(listing_id, booking, id);
     ResponseEntity<BookingModel> response = new ResponseEntity<BookingModel>(m, HttpStatus.OK);
     return response;
@@ -48,7 +71,12 @@ public class BookingController {
 
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
   public ResponseEntity<Boolean> delete(@PathVariable("listing_id") int listing_id,
-      @PathVariable("id") int id) throws EntityNotFoundException {
+      @PathVariable("id") int id, HttpSession session, HttpServletResponse servletResponse)
+      throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     service.delete(listing_id, id);
     ResponseEntity<Boolean> response = new ResponseEntity<Boolean>(true, HttpStatus.NO_CONTENT);
     return response;
