+++ src/main/java/com/practo/jedi/carpool/controller/VehicleController.java
@@ -1,5 +1,8 @@
 package com.practo.jedi.carpool.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,32 +24,59 @@ public class VehicleController {
   private VehicleService service;
 
   @RequestMapping(method = RequestMethod.GET)
-  public Iterable<VehicleModel> list(@PathVariable("user_id") int user_id) {
+  public Iterable<VehicleModel> list(@PathVariable("user_id") int user_id, HttpSession session,
+      HttpServletResponse servletResponse) {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get(user_id);
   }
 
   @RequestMapping(method = RequestMethod.POST)
-  public ResponseEntity<VehicleModel> create(@PathVariable("user_id") int user_id, @RequestBody VehicleModel vehicle) throws EntityNotFoundException {
+  public ResponseEntity<VehicleModel> create(@PathVariable("user_id") int user_id,
+      @RequestBody VehicleModel vehicle, HttpSession session, HttpServletResponse servletResponse)
+      throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     VehicleModel m = service.create(user_id, vehicle);
     ResponseEntity<VehicleModel> response = new ResponseEntity<VehicleModel>(m, HttpStatus.CREATED);
     return response;
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  public VehicleModel get(@PathVariable("user_id") int user_id, @PathVariable("id") int id) throws EntityNotFoundException {
+  public VehicleModel get(@PathVariable("user_id") int user_id, @PathVariable("id") int id,
+      HttpSession session, HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     return service.get(user_id, id);
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-  public ResponseEntity<VehicleModel> update(@PathVariable("user_id") int user_id, @PathVariable("id") int id,
-      @RequestBody VehicleModel vehicle) throws EntityNotFoundException {
+  public ResponseEntity<VehicleModel> update(@PathVariable("user_id") int user_id,
+      @PathVariable("id") int id, @RequestBody VehicleModel vehicle, HttpSession session,
+      HttpServletResponse servletResponse) throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     VehicleModel m = service.update(user_id, vehicle, id);
     ResponseEntity<VehicleModel> response = new ResponseEntity<VehicleModel>(m, HttpStatus.OK);
     return response;
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-  public ResponseEntity<Boolean> delete(@PathVariable("user_id") int user_id, @PathVariable("id") int id) throws EntityNotFoundException {
+  public ResponseEntity<Boolean> delete(@PathVariable("user_id") int user_id,
+      @PathVariable("id") int id, HttpSession session, HttpServletResponse servletResponse)
+      throws EntityNotFoundException {
+    if (session.getAttribute("user") == null) {
+      servletResponse.setStatus(401);
+      return null;
+    }
     service.delete(user_id, id);
     ResponseEntity<Boolean> response = new ResponseEntity<Boolean>(true, HttpStatus.NO_CONTENT);
     return response;
